Create views for all categories, rename component
