+++ AgilProjekt/src/components/QuizComponent.vue
