+++ AgilProjekt/src/views/LanguageView.vue
@@ -4,6 +4,6 @@ import QuizComponent from '../components/QuizComponent.vue'
 
 <template>
   <main>
-    <QuizComponent category="history" />
+    <QuizComponent category="language" />
   </main>
 </template>
